bind to outgoing ip only when specifically requested
we experienced that when binding to 127.0.0.1, no outgoing connections
could be made, and rather than blacklisting all possible ip addresses
that need bind turned off in order for the proxy to work, we instead
add a new command line switch so the user can turn bind mode on when
he sees the need to do so.

the bind functionality was introduced in this commit's parent.