add -w whitelist option
based on the idea of @mimi89999 and later @gallantron.

the whitelist option is the logical conclusion to complement
the -1 (auth-once) option, and i've been in favor of adding
the feature since i first saw the idea.

if the user's ip is known beforehand, one doesn't have to
run a command especially to "unlock" the ip for no-auth use.
one of the few cases where it makes sense to add another
command line option, since we're basically just reusing
an already existing feature.

however there were some technical issues in the original PR
(like using stdbool.h) and not using a list, which made me
want to write the functionality myself. until i forgot.

now there's been a second PR implementing the same
functionality again, but with more complex rules, and as well
without making it possible to process a list of ips.

closes #46
closes #75